Use a simple CSV-based file storage instead of the SQL databases

+++ src/config/mod.rs
@@ -33,7 +33,7 @@ pub struct AppConfig {
     pub blacklisted_domains: Vec<String>,
     pub db_dsn: String,
     pub filters: Vec<ItemFilter>,
-    pub purge_after_days: usize,
+    pub purge_after_days: i64,
     pub smtp: Option<SmtpConfig>,
     pub telegram: Option<TelegramConfig>,
     pub rss_sources: Option<Vec<RssSource>>,
